[KED-2542] Make sure pipelines are lazily loaded (#1060)
