+++ RELEASE.md
@@ -3,6 +3,13 @@
 ## Major features and improvements
 * Kedro plugins can now override built-in CLI commands.
 * Added a `before_command_run` hook for plugins to add extra behaviour before Kedro CLI commands run.
+* `pipelines` from `pipeline_registry.py` and `register_pipeline` hooks are now loaded lazily when they are first accessed, not on startup:
+
+```python
+from kedro.framework.project import pipelines
+
+print(pipelines["__default__"])  # pipeline loading is only triggered here
+```
 
 ## Bug fixes and other changes
 * `TemplatedConfigLoader` now correctly inserts default values when no globals are supplied.
@@ -12,6 +19,7 @@
 * CLI commands from sources with the same name will show under one list in the help screen.
 * The setup of a Kedro project, including adding src to path and configuring settings, is now handled via the `bootstrap_project` method.
 * Invoked `configure_project` if a `package_name` is supplied to `KedroSession.create`. This is added for backward-compatibility purpose to support workflow that creates a `Session` manually. It will only be removed in `0.18.0`.
+* Stopped swallowing up all `ModuleNotFoundError` if `register_pipelines` not found, which will display a more helpful error when a dependency is missing, e.g. [Issue #722](https://github.com/quantumblacklabs/kedro/issues/722).
 
 ## Minor breaking changes to the API
