+++ kedro/framework/cli/hooks/markers.py
@@ -33,5 +33,5 @@ import pluggy
 
 CLI_HOOK_NAMESPACE = "kedro_cli"
 
-cli_hook_spec = pluggy.HookspecMarker(CLI_HOOK_NAMESPACE)
+cli_hook_spec = pluggy.HookspecMarker(CLI_HOOK_NAMESPACE)  # pragma: no cover
 cli_hook_impl = pluggy.HookimplMarker(CLI_HOOK_NAMESPACE)
