+++ kedro/framework/context/context.py
@@ -293,7 +293,7 @@ class KedroContext:
         Returns:
             A dictionary of defined pipelines.
         """
-        return pipelines
+        return dict(pipelines)
 
     def _get_pipeline(  # pylint: disable=no-self-use
         self, name: str = None
